// src/plugins/mg-death/shared/enums.ts
export enum MissionType {
    None = 0,
    Cruise = 1,
    Ram = 2,
    Block = 3,
    GoTo = 4,
    Stop = 5,
    Attack = 6,
    Follow = 7,
    Flee = 8,
    Circle = 9,
    Escort = 12,
    GoToRacing = 14,
    FollowRecording = 15,
    PoliceBehaviour = 16,
    Land = 19,
    LandAndWait = 20,
    Crash = 21,
    PullOver = 22,
    HeliProtect = 23 
};

export enum MissionFlag {
    None = 0,
    AttainRequestedOrientation = 1,
    DontModifyOrientation = 2,
    DontModifyPitch = 4,
    DontModifyThrottle = 8,
    DontModifyRoll = 16,
    LandOnArrival = 32,
    DontDoAvoidance = 64,
    StartEngineImmediately = 128,
    ForceHeightMapAvoidance = 256,
    DontClampProbesToDestination = 512,
    EnableTimeslicingWhenPossible = 1024,
    CircleOppositeDirection = 2048,
    MaintainHeightAboveTerrain = 4096,
    IgnoreHiddenEntitiesDuringLand = 8192,
    DisableAllHeightMapAvoidance = 16384,
    HeightMapOnlyAvoidance = 320,
};